Remove workarounds for pre java8
Motivation:

As we now have a requirement of java11+ we can remove some of the hacks to support older versions then java8

Modifications:

Remove some of the hacks and merge some other code

Result:

Cleanup
