+++ handler/src/test/java/io/netty5/handler/ssl/OpenSslConscryptSslEngineInteropTest.java
@@ -173,7 +173,7 @@ public class OpenSslConscryptSslEngineInteropTest extends ConscryptSslEngineTest
 
     @Override
     protected SSLEngine wrapEngine(SSLEngine engine) {
-        return Java8SslTestUtils.wrapSSLEngineForTesting(engine);
+        return SslTestUtils.wrapSSLEngineForTesting(engine);
     }
 
     @SuppressWarnings("deprecation")
