+++ handler/src/test/java/io/netty5/handler/ssl/OpenSslJdkSslEngineInteroptTest.java
@@ -186,7 +186,7 @@ public class OpenSslJdkSslEngineInteroptTest extends SSLEngineTest {
 
     @Override
     protected SSLEngine wrapEngine(SSLEngine engine) {
-        return Java8SslTestUtils.wrapSSLEngineForTesting(engine);
+        return SslTestUtils.wrapSSLEngineForTesting(engine);
     }
 
     @SuppressWarnings("deprecation")
