+++ handler/src/test/java/io/netty5/handler/ssl/SslContextTrustManagerTest.java
@@ -24,7 +24,7 @@ import java.security.cert.CertificateException;
 import java.security.cert.X509Certificate;
 import java.util.Arrays;
 
-import static io.netty5.handler.ssl.Java8SslTestUtils.loadCertCollection;
+import static io.netty5.handler.ssl.SslTestUtils.loadCertCollection;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.fail;
